+++ src/converse-core.js
@@ -78,6 +78,7 @@
         'converse-dropdown',
         'converse-fullscreen',
         'converse-headline',
+        'converse-http-file-upload',
         'converse-mam',
         'converse-minimize',
         'converse-modal',