+++ CHANGES.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Http-File-Upload
+
+## New Features
+
+- files can now be send via http-file-upload (XEP-0363)
+- mp4 and mp3 files will now be playable directly in chat. 
+
 ## 4.0.0 (Unreleased)
 
 ## UI changes